requirements.txt
requests==2.28.1
psycopg2-binary==2.9.6
pytest==7.2.2
python-dotenv==1.0.0
pandas==1.4.0